+++ Bigfish/test/testcode1.py
@@ -5,4 +5,3 @@ def init():
 
 def handle():
     print(O[0], H[0], L[0], C[0])
-    asd
