+++ Bigfish/test/testcode16.py
@@ -1,0 +1,11 @@
+# -*- coding:utf-8 -*-
+
+
+def init():
+    pass
+
+
+def handle():
+    # 若当前没有持仓，且收盘价小于开盘价
+    num = BarNum - 1
+    print(Close[num])
