+++ Bigfish/test/testcode5.py
@@ -5,5 +5,7 @@ def init():
     pass
 
 
-def handle():
-    print(ATR(3,0))
+def handle(length=5):
+    boll = Bolling(length)
+    if boll is not None:
+        print(boll.ma, boll.std, boll.up, boll.down)
